final edit

+++ WDC-F24/Controllers/ProductController.cs
@@ -1,5 +1,7 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 using WDC_F24.Application.DTOs.Requests;
 using WDC_F24.Application.DTOs.Responses;
 using WDC_F24.Application.Interfaces;
@@ -21,8 +23,7 @@ public class ProductsController : ControllerBase
     }
 
 
-    [HttpGet("GetProduct")]
-    //[ApiExplorerSettings(GroupName = "UserApp")]
+    [HttpGet]
     public async Task<IActionResult> Get()
     {
         if (!ModelState.IsValid)
@@ -59,14 +60,19 @@ public class ProductsController : ControllerBase
         }
 
     }
-
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(Guid id)
     {
+        var userName = _httpContextAccessor.HttpContext?.User.FindFirstValue(ClaimTypes.Name);
+        var UserIdentifier = User.FindFirstValue(ClaimTypes.NameIdentifier);
         if (!ModelState.IsValid)
             return BadRequest(ModelState);
         try
         {
+            if (!Guid.TryParse(UserIdentifier, out var userid))
+                return Unauthorized("Invalid token");
+
             var Result = await _productService.DeleteAsync(id);
             var Response = Result.ToActionResult();
             return Response;
@@ -77,14 +83,19 @@ public class ProductsController : ControllerBase
         }
 
     }
-   
+    [Authorize]
     [HttpPost("AddProduct")]
     public async Task<IActionResult> Add(AddProudctRequestDto product)
     {
+        
+        var UserIdentifier = User.FindFirstValue(ClaimTypes.NameIdentifier);
         if (!ModelState.IsValid)
             return BadRequest(ModelState);
         try
         {
+         
+            if (string.IsNullOrEmpty(UserIdentifier) || !Guid.TryParse(UserIdentifier, out var userId))
+                return Unauthorized("Invalid token or user ID");
             var Result = await _productService.AddAsync(product);
             var Response = Result.ToActionResult();
             return Response;
@@ -94,15 +105,19 @@ public class ProductsController : ControllerBase
             return BadRequest(ex.Message);
         }
 
-    }  
+    }
+    [Authorize]
     [HttpPatch("UpdateProduct")]
-    public async Task<IActionResult> Update(Guid id,AddProudctRequestDto product )
+    public async Task<IActionResult> Update(UpdateProductRequestDto product )
     {
+        var UserIdentifier = User.FindFirstValue(ClaimTypes.NameIdentifier);
         if (!ModelState.IsValid)
             return BadRequest(ModelState);
         try
         {
-            var Result = await _productService.UpdateAsync(product , id);
+            if (string.IsNullOrEmpty(UserIdentifier) || !Guid.TryParse(UserIdentifier, out var userId))
+                return Unauthorized("Invalid token or user ID");
+            var Result = await _productService.UpdateAsync(product);
             var Response = Result.ToActionResult();
             return Response;
         }
